+++ packages/gestalt/src/SearchField.js
@@ -8,6 +8,7 @@ import Icon from './Icon.js';
 
 type Props = {|
   accessibilityLabel: string,
+  autoComplete?: 'current-password' | 'on' | 'off' | 'username',
   id: string,
   onBlur?: ({ event: SyntheticEvent<HTMLInputElement> }) => void,
   onChange: ({
@@ -30,6 +31,12 @@ type State = {|
 export default class SearchField extends React.Component<Props, State> {
   static propTypes = {
     accessibilityLabel: PropTypes.string.isRequired,
+    autoComplete: PropTypes.oneOf([
+      'current-password',
+      'on',
+      'off',
+      'username',
+    ]),
     id: PropTypes.string.isRequired,
     onBlur: PropTypes.func,
     onChange: PropTypes.func.isRequired,
@@ -81,7 +88,13 @@ export default class SearchField extends React.Component<Props, State> {
   };
 
   render() {
-    const { accessibilityLabel, id, placeholder, value } = this.props;
+    const {
+      accessibilityLabel,
+      autoComplete,
+      id,
+      placeholder,
+      value,
+    } = this.props;
 
     // This mirrors the built in browser behavior. If there's a value, show the
     // clear button if you're hovering or if you've focused on the field
@@ -116,6 +129,7 @@ export default class SearchField extends React.Component<Props, State> {
         </Box>
         <input
           aria-label={accessibilityLabel}
+          autoComplete={autoComplete}
           className={styles.input}
           id={id}
           onChange={this.handleChange}